requirements.txt
# Cambrionix virtualenv dependencies

pyserial==3.5.0
enum34==1.1.10
six==1.14.0
pyyaml==5.3.1
colorlog==4.2.1
pylint==1.9.5
pytest==4.6.9
pytest-cov==2.8.1
serial_mock==0.1
ipython==5.8.0
